# api/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API Gateway of Personalized Medicine and Health Monitoring Platform",
        "title": "Personalized Medicine and Health Monitoring Platform",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/customer/wearable": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Creates a new wearable data record for a user",
                "tags": [
                    "wearable"
                ],
                "summary": "Adds new wearable data",
                "parameters": [
                    {
                        "description": "Wearable data",
                        "name": "wearable_data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/health_analytics.AddWearableDataRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/health_analytics.AddWearableDataResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error while processing request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/customer/wearable/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieves wearable data by ID",
                "tags": [
                    "wearable"
                ],
                "summary": "Gets wearable data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Wearable data ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/health_analytics.GetWearableDataResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid data ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error while processing request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Updates existing wearable data",
                "tags": [
                    "wearable"
                ],
                "summary": "Updates wearable data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Wearable data ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated wearable data",
                        "name": "wearable_data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/health_analytics.UpdateWearableDataRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/health_analytics.UpdateWearableDataResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error while processing request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Deletes wearable data by ID",
                "tags": [
                    "wearable"
                ],
                "summary": "Deletes wearable data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Wearable data ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/health_analytics.DeleteWearableDataResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid data ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error while processing request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "health_analytics.AddWearableDataRequest": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/health_analytics.WearableData"
                }
            }
        },
        "health_analytics.AddWearableDataResponse": {
            "type": "object",
            "properties": {
                "data_id": {
                    "type": "string"
                }
            }
        },
...
        "health_analytics.DeleteWearableDataResponse": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "health_analytics.GetWearableDataResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/health_analytics.WearableData"
                }
            }
        },
...
        "health_analytics.UpdateWearableDataRequest": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/health_analytics.WearableData"
                }
            }
        },
        "health_analytics.UpdateWearableDataResponse": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean"
                }
            }
        },
        "health_analytics.WearableData": {
            "type": "object",
            "properties": {
                "data_type": {
                    "type": "string"
                },
                "data_value": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "device_type": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "recorded_timestamp": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}